^r Extract credential and SSH rendering from render_injection_bundle.go, move-only (gofmt, vet, build, package unit+fuzz tests pass; declaration set verified identical; internal restructure, no behavior change)
Step 2 of 3 for the injection god-file split: renderCredentials with the
Gemini env/JSON credential validators moves verbatim to
render_credentials.go, and renderSSH with the known-hosts and SSH config
safety checks to render_ssh.go.
