+++ internal/injection/render_injection_bundle.go
@@ -23,7 +23,6 @@ import (
 	"github.com/omkhar/workcell/internal/injectionpolicy"
 	"github.com/omkhar/workcell/internal/pathutil"
 	"github.com/omkhar/workcell/internal/providerid"
-	"github.com/omkhar/workcell/internal/tomlsubset"
 )
 
 const (
@@ -339,245 +338,6 @@ func renderCopies(policy map[string]any, outputRoot, policyDir Path, agent, mode
 			"dir_mode":       dirMode,
 			"classification": classification,
 		})
-	}
-	return rendered, nil
-}
-
-func renderSSH(policy map[string]any, outputRoot, policyDir Path, agent, mode string) (map[string]any, error) {
-	raw := policy["ssh"]
-	if raw == nil {
-		return map[string]any{}, nil
-	}
-	ssh, ok := raw.(map[string]any)
-	if !ok {
-		return nil, errors.New("ssh must be a TOML table")
-	}
-	if err := validateAllowedKeys(ssh, mapKeysSet([]string{"enabled", "config", "known_hosts", "identities", "providers", "modes", "allow_unsafe_config"}), "ssh"); err != nil {
-		return nil, err
-	}
-	enabledRaw, hasEnabled := ssh["enabled"]
-	hasMaterial := false
-	for _, key := range []string{"config", "known_hosts", "identities"} {
-		if _, ok := ssh[key]; ok {
-			hasMaterial = true
-		}
-	}
-	if hasEnabled {
-		enabled, ok := enabledRaw.(bool)
-		if !ok {
-			return nil, errors.New("ssh.enabled must be a boolean when specified")
-		}
-		if !enabled {
-			return map[string]any{}, nil
-		}
-	}
-	if !hasEnabled && !hasMaterial {
-		return map[string]any{}, nil
-	}
-	ok, err := selectedFor(ssh["providers"], agent, "ssh.providers", supportedAgents)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return map[string]any{}, nil
-	}
-	ok, err = selectedFor(ssh["modes"], mode, "ssh.modes", supportedModes)
-	if err != nil {
-		return nil, err
-	}
-	if !ok {
-		return map[string]any{}, nil
-	}
-	rendered := map[string]any{
-		"identities": []map[string]any{},
-	}
-	allowUnsafeRaw, hasAllowUnsafe := ssh["allow_unsafe_config"]
-	allowUnsafeConfig := false
-	if hasAllowUnsafe {
-		val, ok := allowUnsafeRaw.(bool)
-		if !ok {
-			return nil, errors.New("ssh.allow_unsafe_config must be a boolean when specified")
-		}
-		allowUnsafeConfig = val
-	}
-	configRaw, hasConfig := ssh["config"]
-	if !hasConfig || configRaw == nil {
-		rendered["config_assurance"] = "no-config"
-	} else if allowUnsafeConfig {
-		rendered["config_assurance"] = "lower-assurance-unsafe-config"
-	} else {
-		rendered["config_assurance"] = "safe"
-	}
-	if hasConfig && configRaw != nil {
-		source, err := validateSourcePath(configRaw, "ssh.config", policyDir)
-		if err != nil {
-			return nil, err
-		}
-		if _, err := validateSecretFile(source, "ssh.config"); err != nil {
-			return nil, err
-		}
-		if err := validateSSHConfigSafety(source, allowUnsafeConfig); err != nil {
-			return nil, err
-		}
-		rendered["config"] = directMountEntry(source, directMountRoot+"/ssh/config")
-	}
-	knownHostsRaw, hasKnownHosts := ssh["known_hosts"]
-	if hasKnownHosts && knownHostsRaw != nil {
-		source, err := validateSourcePath(knownHostsRaw, "ssh.known_hosts", policyDir)
-		if err != nil {
-			return nil, err
-		}
-		if _, err := validateKnownHostsFile(source, "ssh.known_hosts"); err != nil {
-			return nil, err
-		}
-		rendered["known_hosts"] = directMountEntry(source, directMountRoot+"/ssh/known_hosts")
-	}
-	identitiesRaw, hasIdentities := ssh["identities"]
-	identities := []any{}
-	if hasIdentities {
-		if identitiesRaw == nil {
-			identities = []any{}
-		} else {
-			var err error
-			identities, err = anySlice(identitiesRaw, "ssh.identities")
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	renderedIdentities := make([]map[string]any, 0, len(identities))
-	seenIdentityTargets := map[string]struct{}{}
-	for index, rawIdentity := range identities {
-		source, err := validateSourcePath(rawIdentity, fmt.Sprintf("ssh.identities[%d]", index), policyDir)
-		if err != nil {
-			return nil, err
-		}
-		if _, err := validateSecretFile(source, fmt.Sprintf("ssh.identities[%d]", index)); err != nil {
-			return nil, err
-		}
-		if _, reserved := reservedSSHFilnames[source.Base()]; reserved {
-			return nil, fmt.Errorf("ssh.identities[%d] basename collides with a reserved SSH file: %s", index, source.Base())
-		}
-		if _, exists := seenIdentityTargets[source.Base()]; exists {
-			return nil, fmt.Errorf("ssh.identities contains duplicate target basename: %s", source.Base())
-		}
-		seenIdentityTargets[source.Base()] = struct{}{}
-		renderedIdentities = append(renderedIdentities, map[string]any{
-			"source":      source.String(),
-			"mount_path":  directMountRoot + "/ssh/identity-" + strconv.Itoa(index),
-			"target_name": source.Base(),
-		})
-	}
-	rendered["identities"] = renderedIdentities
-	return rendered, nil
-}
-
-func renderCredentials(policy map[string]any, policyDir Path, agent, mode string) (map[string]map[string]string, error) {
-	raw := policy["credentials"]
-	if raw == nil {
-		return map[string]map[string]string{}, nil
-	}
-	credentials, ok := raw.(map[string]any)
-	if !ok {
-		return nil, errors.New("credentials must be a TOML table")
-	}
-	if err := validateAllowedKeys(credentials, mapKeysSet([]string{
-		"codex_auth",
-		"claude_auth",
-		"claude_api_key",
-		"claude_mcp",
-		"gemini_env",
-		"gemini_oauth",
-		"gemini_projects",
-		"gcloud_adc",
-		"github_hosts",
-		"github_config",
-	}), "credentials"); err != nil {
-		return nil, err
-	}
-
-	relevant := map[string]struct{}{}
-	for key := range sharedCredentialKeys {
-		relevant[key] = struct{}{}
-	}
-	if scoped, ok := agentScopedCredentialKeys[agent]; ok {
-		for key := range scoped {
-			relevant[key] = struct{}{}
-		}
-	}
-
-	rendered := map[string]map[string]string{}
-	for _, key := range sortedKeys(relevant) {
-		rawValue, ok := credentials[key]
-		if !ok || rawValue == nil {
-			continue
-		}
-
-		var providers any
-		var modes any
-		sourceRaw := rawValue
-		entry, isTable := rawValue.(map[string]any)
-		if isTable {
-			if err := validateAllowedKeys(entry, mapKeysSet([]string{"source", "providers", "modes"}), "credentials."+key); err != nil {
-				return nil, err
-			}
-			sourceRaw = entry["source"]
-			providers = entry["providers"]
-			modes = entry["modes"]
-		} else if _, ok := rawValue.(string); !ok {
-			return nil, fmt.Errorf("credentials.%s must be a string path or a table", key)
-		}
-
-		if _, shared := sharedCredentialKeys[key]; shared {
-			if !isTable || providers == nil {
-				return nil, fmt.Errorf("credentials.%s.providers is required so shared GitHub credentials stay least-privilege", key)
-			}
-		}
-		ok, err := selectedFor(providers, agent, "credentials."+key+".providers", supportedAgents)
-		if err != nil {
-			return nil, err
-		}
-		if !ok {
-			continue
-		}
-		ok, err = selectedFor(modes, mode, "credentials."+key+".modes", supportedModes)
-		if err != nil {
-			return nil, err
-		}
-		if !ok {
-			continue
-		}
-
-		source, err := validateSourcePath(sourceRaw, "credentials."+key, policyDir)
-		if err != nil {
-			return nil, err
-		}
-		source, err = validateSecretFile(source, "credentials."+key)
-		if err != nil {
-			return nil, err
-		}
-		switch key {
-		case "gemini_env":
-			if _, err := validateGeminiEnvFile(source); err != nil {
-				return nil, err
-			}
-		case "gemini_oauth":
-			if _, err := validateJSONObjFile(source, "credentials.gemini_oauth"); err != nil {
-				return nil, err
-			}
-		case "gcloud_adc":
-			if err := validateGcloudADCFile(source, "credentials.gcloud_adc"); err != nil {
-				return nil, err
-			}
-		case "gemini_projects":
-			if err := validateGeminiProjectsFile(source, "credentials.gemini_projects"); err != nil {
-				return nil, err
-			}
-		}
-		rendered[key] = map[string]string{
-			"source":     source.String(),
-			"mount_path": credentialContainerPaths[key],
-		}
 	}
 	return rendered, nil
 }
@@ -908,264 +668,6 @@ func validateSecretTree(source Path, label string) error {
 		}
 		return requireSecretOwnerOnly(child, label)
 	})
-}
-
-func validateKnownHostsFile(source Path, label string) (Path, error) {
-	if err := requireNoSymlink(source, label); err != nil {
-		return Path(""), err
-	}
-	info, err := os.Stat(source.String())
-	if err != nil {
-		return Path(""), err
-	}
-	if !info.Mode().IsRegular() {
-		return Path(""), fmt.Errorf("%s must point at a file: %s", label, source)
-	}
-	if info.Mode().Perm()&0o022 != 0 {
-		return Path(""), fmt.Errorf("%s must not be group/world-writable: %s", label, source)
-	}
-	return source, nil
-}
-
-func parseSSHDirective(line string) (string, string, bool) {
-	stripped := strings.TrimSpace(line)
-	if stripped == "" || strings.HasPrefix(stripped, "#") {
-		return "", "", false
-	}
-	parts := strings.Fields(stripped)
-	directive := strings.ToLower(parts[0])
-	remainder := ""
-	if len(parts) > 1 {
-		remainder = strings.Join(parts[1:], " ")
-	}
-	return directive, remainder, true
-}
-
-func validateSSHConfigSafety(source Path, allowUnsafe bool) error {
-	if allowUnsafe {
-		return nil
-	}
-	data, err := os.ReadFile(source.String())
-	if err != nil {
-		return err
-	}
-	lines := strings.Split(string(data), "\n")
-	for i, line := range lines {
-		directive, remainder, ok := parseSSHDirective(line)
-		if !ok {
-			continue
-		}
-		if _, risky := riskySSHDirectives[directive]; risky {
-			return fmt.Errorf("ssh.config contains unsafe directive %q at line %d; set ssh.allow_unsafe_config = true only when you explicitly accept lower assurance", directive, i+1)
-		}
-		if directive == "match" && strings.Contains(" "+strings.ToLower(remainder)+" ", " exec ") {
-			return fmt.Errorf("ssh.config contains unsafe Match exec at line %d; set ssh.allow_unsafe_config = true only when you explicitly accept lower assurance", i+1)
-		}
-	}
-	return nil
-}
-
-func validateGeminiEnvFile(source Path) (map[string]any, error) {
-	values, err := parseSimpleEnvFile(source)
-	if err != nil {
-		return nil, err
-	}
-	for key, value := range values {
-		if _, ok := geminiSupportedEnvKeys[key]; !ok {
-			return nil, fmt.Errorf("unsupported key in Gemini auth env file %s: %s", source, key)
-		}
-		if key != "GOOGLE_GENAI_USE_GCA" && key != "GOOGLE_GENAI_USE_VERTEXAI" && strings.TrimSpace(value) == "" {
-			return nil, fmt.Errorf("gemini auth env file %s sets %s but leaves it empty", source, key)
-		}
-	}
-
-	gcaEnabled, err := parseEnvBooleanValue(values, source, "GOOGLE_GENAI_USE_GCA")
-	if err != nil {
-		return nil, err
-	}
-	vertexEnabled, err := parseEnvBooleanValue(values, source, "GOOGLE_GENAI_USE_VERTEXAI")
-	if err != nil {
-		return nil, err
-	}
-	geminiAPIKey := strings.TrimSpace(values["GEMINI_API_KEY"])
-	googleAPIKey := strings.TrimSpace(values["GOOGLE_API_KEY"])
-	hasProject := false
-	for _, key := range geminiProjectKeys {
-		if strings.TrimSpace(values[key]) != "" {
-			hasProject = true
-			break
-		}
-	}
-	hasLocation := false
-	for _, key := range geminiVertexLocationKeys {
-		if strings.TrimSpace(values[key]) != "" {
-			hasLocation = true
-			break
-		}
-	}
-	if gcaEnabled && vertexEnabled {
-		return nil, fmt.Errorf("gemini auth env file %s enables both GOOGLE_GENAI_USE_GCA and GOOGLE_GENAI_USE_VERTEXAI; choose exactly one auth selector", source)
-	}
-	if hasLocation && !hasProject {
-		return nil, fmt.Errorf("gemini auth env file %s sets a Google Cloud location without a project", source)
-	}
-	endpoints := map[string]struct{}{}
-	if vertexEnabled {
-		if googleAPIKey != "" || (hasProject && hasLocation) {
-			endpoints[vertexEndpoint] = struct{}{}
-			for _, key := range geminiVertexLocationKeys {
-				location := normalizeVertexLocation(values[key])
-				if location != "" {
-					endpoints[location+"-aiplatform.googleapis.com:443"] = struct{}{}
-				}
-			}
-			return map[string]any{
-				"selected_auth_type": "vertex-ai",
-				"extra_endpoints":    sortedKeys(endpoints),
-			}, nil
-		}
-		return nil, fmt.Errorf("gemini auth env file %s enables GOOGLE_GENAI_USE_VERTEXAI=true without either GOOGLE_API_KEY or both GOOGLE_CLOUD_PROJECT and GOOGLE_CLOUD_LOCATION", source)
-	}
-	if googleAPIKey != "" {
-		return nil, fmt.Errorf("gemini auth env file %s sets GOOGLE_API_KEY without GOOGLE_GENAI_USE_VERTEXAI=true", source)
-	}
-	if gcaEnabled {
-		return map[string]any{
-			"selected_auth_type": "oauth-personal",
-			"extra_endpoints":    append([]string{}, googleAuthEndpoints...),
-		}, nil
-	}
-	if geminiAPIKey != "" {
-		return map[string]any{
-			"selected_auth_type": "gemini-api-key",
-			"extra_endpoints":    []string{},
-		}, nil
-	}
-	return nil, fmt.Errorf("gemini auth env file %s does not configure a supported Gemini auth mode; use GEMINI_API_KEY, GOOGLE_GENAI_USE_GCA=true, or GOOGLE_GENAI_USE_VERTEXAI=true with GOOGLE_API_KEY or both GOOGLE_CLOUD_PROJECT and GOOGLE_CLOUD_LOCATION", source)
-}
-
-func parseSimpleEnvFile(source Path) (map[string]string, error) {
-	values := map[string]string{}
-	data, err := os.ReadFile(source.String())
-	if err != nil {
-		return nil, err
-	}
-	for _, rawLine := range strings.Split(string(data), "\n") {
-		line := strings.TrimSpace(rawLine)
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		if strings.HasPrefix(line, "export ") {
-			line = strings.TrimSpace(line[len("export "):])
-		}
-		idx := strings.Index(line, "=")
-		if idx < 0 {
-			return nil, fmt.Errorf("malformed Gemini auth env file %s: %s; use KEY=value assignments", source, strings.TrimSpace(rawLine))
-		}
-		key := strings.TrimSpace(line[:idx])
-		value := tomlsubset.StripComment(line[idx+1:])
-		if _, exists := values[key]; exists {
-			return nil, fmt.Errorf("gemini auth env file %s configures %s more than once", source, key)
-		}
-		if value != "" && (value[0] == '\'' || value[0] == '"') {
-			if len(value) < 2 || value[len(value)-1] != value[0] {
-				return nil, fmt.Errorf("malformed Gemini auth env file %s: %s has an unterminated quoted value", source, key)
-			}
-			if value[0] == '"' {
-				parsed, err := strconv.Unquote(value)
-				if err != nil {
-					return nil, fmt.Errorf("malformed Gemini auth env file %s: %s has an invalid double-quoted value (%v)", source, key, err)
-				}
-				value = parsed
-			} else {
-				value = value[1 : len(value)-1]
-			}
-		} else if value != "" && (strings.HasSuffix(value, "'") || strings.HasSuffix(value, "\"")) {
-			return nil, fmt.Errorf("malformed Gemini auth env file %s: %s has an unmatched trailing quote", source, key)
-		}
-		values[key] = value
-	}
-	return values, nil
-}
-
-func parseEnvBooleanValue(values map[string]string, source Path, key string) (bool, error) {
-	raw, ok := values[key]
-	if !ok {
-		return false, nil
-	}
-	normalized := strings.ToLower(strings.TrimSpace(raw))
-	if normalized != "true" && normalized != "false" {
-		return false, fmt.Errorf("invalid boolean in Gemini auth env file %s: %s=%s; use true or false", source, key, raw)
-	}
-	return normalized == "true", nil
-}
-
-func normalizeVertexLocation(value string) string {
-	candidate := strings.ToLower(strings.TrimSpace(value))
-	if candidate == "" {
-		return ""
-	}
-	for _, r := range candidate {
-		if !(r >= 'a' && r <= 'z' || r >= '0' && r <= '9' || r == '-') {
-			return ""
-		}
-	}
-	if strings.HasPrefix(candidate, "-") || strings.HasSuffix(candidate, "-") || strings.Contains(candidate, "--") {
-		return ""
-	}
-	return candidate
-}
-
-func validateJSONObjFile(source Path, label string) (map[string]any, error) {
-	data, err := os.ReadFile(source.String())
-	if err != nil {
-		return nil, err
-	}
-	var parsed any
-	if err := json.Unmarshal(data, &parsed); err != nil {
-		return nil, fmt.Errorf("%s must contain valid JSON: %s (%s)", label, source, err.Error())
-	}
-	object, ok := parsed.(map[string]any)
-	if !ok {
-		return nil, fmt.Errorf("%s must contain a JSON object: %s", label, source)
-	}
-	return object, nil
-}
-
-func validateGcloudADCFile(source Path, label string) error {
-	parsed, err := validateJSONObjFile(source, label)
-	if err != nil {
-		return err
-	}
-	adcType, ok := parsed["type"].(string)
-	if !ok || strings.TrimSpace(adcType) == "" {
-		return fmt.Errorf("%s must contain a non-empty JSON string field: type", label)
-	}
-	return nil
-}
-
-func validateGeminiProjectsFile(source Path, label string) error {
-	parsed, err := validateJSONObjFile(source, label)
-	if err != nil {
-		return err
-	}
-	if _, ok := parsed["projects"].(map[string]any); !ok {
-		return fmt.Errorf("%s must contain a JSON object with an object-valued projects field", label)
-	}
-	return nil
-}
-
-func deriveCredentialExtraEndpoints(renderedCredentials map[string]map[string]string) []string {
-	endpoints := map[string]struct{}{}
-	if geminiEnv, ok := renderedCredentials["gemini_env"]; ok {
-		metadata, err := validateGeminiEnvFile(Path(geminiEnv["source"]))
-		if err == nil {
-			for _, endpoint := range metadata["extra_endpoints"].([]string) {
-				endpoints[endpoint] = struct{}{}
-			}
-		}
-	}
-	return sortedKeys(endpoints)
 }
 
 func effectivePolicySHA256(
